+++ backend/__tests__/route/podcast.episode.test.ts
@@ -1,0 +1,231 @@
+import request from "supertest"
+import { NextFunction, Request, Response } from "express"
+import { afterEach, beforeEach, describe, expect, test, vi } from "vitest"
+import { PODCAST_BY_FEED_ID_75075 } from "../mocks/podcast.js"
+import { setupApp } from "../../index.js"
+import { getFrontendOrigin } from "../cors/origin.js"
+import { getSanitizedHtmlText } from "../../api/dom/htmlSanitize.js"
+
+function getMockMiddleware() {
+  return (request: Request, response: Response, next: NextFunction) => next()
+}
+
+function mockRateLimiters() {
+  vi.mock("../../middleware/rateLimiter.js", () => {
+    return {
+      default: {
+        getTrendingPodcastLimiter: getMockMiddleware(),
+        getPodcastEpisodesLimiter: getMockMiddleware(),
+      },
+    }
+  })
+}
+
+describe("GET /api/podcast/episodes", () => {
+  const expectedOrigin = getFrontendOrigin() || ""
+
+  function getExpectedEpisodeData(episodeItems: any[]) {
+    // convert PodcastIndex API episode "items" response array to expected /api/podcast/episodes json response for "data"
+    return episodeItems.map((episode) => {
+      return {
+        id: episode.id,
+        feedId: episode.feedId,
+        feedUrl: episode.feedUrl,
+        title: episode.title,
+        description: getSanitizedHtmlText(episode.description || ""),
+        contentUrl: episode.enclosureUrl, // url link to episode file
+        contentType: episode.enclosureType, // Content-Type of the episode file (e.g. mp3 => "audio\/mpeg")
+        contentSizeInBytes: episode.enclosureLength,
+        durationInSeconds: episode.duration,
+        datePublished: episode.datePublished, // unix epoch time in seconds
+        isExplicit: episode.explicit === 1, // Not explicit = 0. Explicit = 1
+        episodeType: episode.episodeType, // type of episode. May be null for "liveItem"
+        episodeNumber: episode.episode,
+        seasonNumber: episode.season,
+        image: episode.image || episode.feedImage,
+        language: episode.feedLanguage,
+        people: episode.persons,
+        externalWebsiteUrl: episode.link,
+        transcripts: episode.transcripts,
+        isActiveFeed: episode.feedDead !== 0,
+      }
+    })
+  }
+
+  beforeEach(() => {
+    mockRateLimiters()
+  })
+
+  afterEach(() => {
+    vi.restoreAllMocks()
+  })
+
+  describe("invalid parameters", () => {
+    describe("id parameter", () => {
+      test("should return status 400 for missing id parameter", async () => {
+        const limit = "10"
+        const app = setupApp()
+        const response = await request(app)
+          .get(`/api/podcast/episodes?limit=${limit}`)
+          .set("Origin", expectedOrigin)
+        expect(response.status).toBe(400)
+        expect(response.body).toEqual(
+          expect.objectContaining({
+            errors: expect.arrayContaining(["'id' should be present"]),
+          })
+        )
+      })
+
+      test("should return status 400 for id parameter of empty string", async () => {
+        const limit = "10"
+        const app = setupApp()
+        const response = await request(app)
+          .get(`/api/podcast/episodes?id=&limit=${limit}`)
+          .set("Origin", expectedOrigin)
+        expect(response.status).toBe(400)
+        expect(response.body).toEqual(
+          expect.objectContaining({
+            errors: expect.arrayContaining(["'id' should be present"]),
+          })
+        )
+      })
+    })
+
+    describe("limit parameter", () => {
+      test("should return status 400 for limit parameter of non numeric value", async () => {
+        const podcastId = "75075"
+        const limit = "a"
+        const app = setupApp()
+        const response = await request(app)
+          .get(`/api/podcast/episodes?id=${podcastId}&limit=${limit}`)
+          .set("Origin", expectedOrigin)
+        expect(response.status).toBe(400)
+        expect(response.body).toEqual(
+          expect.objectContaining({
+            errors: expect.arrayContaining([
+              "'limit' should be between 1 and 100",
+            ]),
+          })
+        )
+      })
+
+      test("should return status 400 for limit parameter of negative value", async () => {
+        const podcastId = "75075"
+        const limit = "-1"
+        const app = setupApp()
+        const response = await request(app)
+          .get(`/api/podcast/episodes?id=${podcastId}&limit=${limit}`)
+          .set("Origin", expectedOrigin)
+        expect(response.status).toBe(400)
+        expect(response.body).toEqual(
+          expect.objectContaining({
+            errors: expect.arrayContaining([
+              "'limit' should be between 1 and 100",
+            ]),
+          })
+        )
+      })
+
+      test("should return status 400 for limit parameter of zero", async () => {
+        const podcastId = "75075"
+        const limit = "0"
+        const app = setupApp()
+        const response = await request(app)
+          .get(`/api/podcast/episodes?id=${podcastId}&limit=${limit}`)
+          .set("Origin", expectedOrigin)
+        expect(response.status).toBe(400)
+        expect(response.body).toEqual(
+          expect.objectContaining({
+            errors: expect.arrayContaining([
+              "'limit' should be between 1 and 100",
+            ]),
+          })
+        )
+      })
+
+      test("should return status 400 for limit parameter of 101", async () => {
+        const podcastId = "75075"
+        const limit = "101"
+        const app = setupApp()
+        const response = await request(app)
+          .get(`/api/podcast/episodes?id=${podcastId}&limit=${limit}`)
+          .set("Origin", expectedOrigin)
+        expect(response.status).toBe(400)
+        expect(response.body).toEqual(
+          expect.objectContaining({
+            errors: expect.arrayContaining([
+              "'limit' should be between 1 and 100",
+            ]),
+          })
+        )
+      })
+    })
+  })
+
+  test("should specify response content type header of application/json", async () => {
+    const podcastId = "75075"
+    const limit = "10"
+    const app = setupApp()
+    const response = await request(app)
+      .get(`/api/podcast/episodes?id=${podcastId}&limit=${limit}`)
+      .set("Origin", expectedOrigin)
+    expect(response.headers["content-type"]).toEqual(
+      expect.stringContaining("application/json")
+    )
+  })
+
+  describe("limit parameter", () => {
+    test("should return 10 episodes based on a podcast id (PodcastIndex Feed Id)", async () => {
+      const podcastId = "75075"
+      const limit = "10"
+      const app = setupApp()
+      const response = await request(app)
+        .get(`/api/podcast/episodes?id=${podcastId}&limit=${limit}`)
+        .set("Origin", expectedOrigin)
+      expect(response.status).toBe(200)
+      expect(response.body).toEqual(
+        expect.objectContaining({
+          count: 10,
+          data: getExpectedEpisodeData(
+            PODCAST_BY_FEED_ID_75075.items.slice(0, 10)
+          ),
+        })
+      )
+    })
+
+    test("should return 1 episode based on a podcast id (PodcastIndex Feed Id)", async () => {
+      const podcastId = "75075"
+      const limit = "1"
+      const app = setupApp()
+      const response = await request(app)
+        .get(`/api/podcast/episodes?id=${podcastId}&limit=${limit}`)
+        .set("Origin", expectedOrigin)
+      expect(response.status).toBe(200)
+      expect(response.body).toEqual(
+        expect.objectContaining({
+          count: 1,
+          data: getExpectedEpisodeData(
+            PODCAST_BY_FEED_ID_75075.items.slice(0, 1)
+          ),
+        })
+      )
+    })
+
+    test("should return 10 episodes by default if no limit parameter is given", async () => {
+      const podcastId = "75075"
+      const app = setupApp()
+      const response = await request(app)
+        .get(`/api/podcast/episodes?id=${podcastId}`)
+        .set("Origin", expectedOrigin)
+      expect(response.status).toBe(200)
+      expect(response.body).toEqual(
+        expect.objectContaining({
+          count: 10,
+          data: getExpectedEpisodeData(
+            PODCAST_BY_FEED_ID_75075.items.slice(0, 10)
+          ),
+        })
+      )
+    })
+  })
+})
